+++ src/features/logging/instrument/index.js
@@ -4,8 +4,8 @@ import { bufferLog } from '../shared/utils'
 
 export class Instrument extends InstrumentBase {
   static featureName = FEATURE_NAME
-  constructor (thisAgent, auto = true) {
-    super(thisAgent, FEATURE_NAME, auto)
+  constructor (agentRef, auto = true) {
+    super(agentRef, FEATURE_NAME, auto)
 
     const instanceEE = this.ee
     /** emitted by wrap-logger function */
@@ -13,7 +13,7 @@ export class Instrument extends InstrumentBase {
       const { level, customAttributes } = this
       bufferLog(instanceEE, message, customAttributes, level)
     })
-    this.importAggregator(thisAgent)
+    this.importAggregator(agentRef)
   }
 }