+++ src/features/page_view_event/instrument/index.js
@@ -3,10 +3,10 @@ import * as CONSTANTS from '../constants'
 
 export class Instrument extends InstrumentBase {
   static featureName = CONSTANTS.FEATURE_NAME
-  constructor (thisAgent, auto = true) {
-    super(thisAgent, CONSTANTS.FEATURE_NAME, auto)
+  constructor (agentRef, auto = true) {
+    super(agentRef, CONSTANTS.FEATURE_NAME, auto)
 
-    this.importAggregator(thisAgent)
+    this.importAggregator(agentRef)
   }
 }