+++ src/features/utils/feature-base.js
@@ -1,4 +1,5 @@
 import { ee } from '../../common/event-emitter/contextual-ee'
+import { deregisterDrain } from '../../common/drain/drain'
 
 export class FeatureBase {
   constructor (agentIdentifier, featureName) {
@@ -14,5 +15,9 @@ export class FeatureBase {
      * @type {boolean}
      */
     this.blocked = false
+  }
+
+  deregisterDrain () {
+    deregisterDrain(this.agentIdentifier, this.featureName)
   }
 }