+++ tests/specs/util/helpers.js
@@ -14,7 +14,7 @@ export const RRWEB_EVENT_TYPES = {
   Custom: 5
 }
 
-export function testExpectedReplay ({ data, session, hasMeta, hasSnapshot, hasError, isFirstChunk, contentEncoding, decompressedBytes, appId, entityGuid, harvestId }) {
+export function testExpectedReplay ({ data, session, hasMeta, hasSnapshot, hasError, isFirstChunk, contentEncoding, decompressedBytes, appId, entityGuid, harvestId, currentUrl }) {
   expect(data.query).toMatchObject({
     browser_monitoring_key: expect.any(String),
     type: 'SessionReplay',
@@ -41,7 +41,8 @@ export function testExpectedReplay ({ data, session, hasMeta, hasSnapshot, hasEr
     isFirstChunk: isFirstChunk || expect.any(Boolean),
     decompressedBytes: decompressedBytes || expect.any(Number),
     'rrweb.version': expect.any(String),
-    inlinedAllStylesheets: expect.any(Boolean)
+    inlinedAllStylesheets: expect.any(Boolean),
+    ...(currentUrl && { currentUrl })
   })
 
   expect(data.body).toEqual(expect.any(Array))
@@ -57,7 +58,8 @@ export function testExpectedTrace ({
   session,
   ptid,
   harvestId,
-  entityGuid
+  entityGuid,
+  currentUrl
 }) {
   expect(data.query).toMatchObject({
     browser_monitoring_key: expect.any(String),
@@ -77,6 +79,7 @@ export function testExpectedTrace ({
     'trace.nodes': nodeCount || expect.any(Number),
     ptid: ptid || expect.anything(),
     session: session || expect.any(String),
+    ...(currentUrl && { currentUrl }),
     // optional attrs here
     ...(firstSessionHarvest && { firstSessionHarvest }),
     ...(hasReplay && { hasReplay })