+++ tests/unit/common/harvest/harvest-scheduler.test.js
@@ -36,15 +36,15 @@ describe('unload', () => {
     expect(subscribeToEOL).toHaveBeenCalledWith(expect.any(Function))
   })
 
-  test('should run onUnload callback when started', () => {
-    harvestSchedulerInstance.opts.onUnload = jest.fn()
+  test('should run onUnload callback on EoL when provided', () => {
+    harvestSchedulerInstance = new HarvestScheduler(undefined, { onUnload: jest.fn() })
 
-    eolSubscribeFn()
+    for (const arr of jest.mocked(subscribeToEOL).mock.calls) arr[0]()
 
     expect(harvestSchedulerInstance.opts.onUnload).toHaveBeenCalledTimes(1)
   })
 
-  test('should run harvest when started and not aborted', () => {
+  test('should run harvest on EoL if not aborted', () => {
     harvestSchedulerInstance.aborted = false
 
     eolSubscribeFn()